Add --max-shown auto mode via negative value
Changes made:
- refactor MaxShown flag to enum { Unset, Auto, Count(usize) } with resolve(top_count) -> Option<usize> and is_set() helpers
- widen cli arg max_shown from Option<usize> to Option<i64> with allow_hyphen_values; config_file mirrors the type widening
- negative cli/config value resolves to MaxShown::Auto; nonneg resolves to Count(n)
- in tree() compute top-level entry count (single-arg dir: content.len(); else metas.len(); single file: 1) and resolve MaxShown into an Option<usize> cap once at entry
- thread the resolved cap explicitly through tree_columns() and inner_display_tree() instead of reading flags.max_shown inside; recursive calls propagate the same cap
- --tree-columns gate uses !is_set() so Auto satisfies the precondition
- add 8 flag tests (zero / negative-cli / large-negative / negative-config / is_set / resolve trio) and 2 display tests (auto truncates subdirs to top-level count; auto satisfies tree-columns gate)

Notes:
- auto cap is resolved once at tree() entry and threaded down; not recomputed per recursion. single-arg dir uses content.len() so user's mental model ("what I see at the top") holds
- max-shown 0 still works (Count(0), truncates everything to a summary line); Auto only produces 0 when the top-level dir is empty, where truncation is a no-op anyway

---

Tested:
- cargo test: 428 passing (8 new flag + 2 new display), 1 pre-existing icon failure unchanged
- smoke: lsd --tree --max-shown -1 ./src truncates every depth to 15 (src has 15 entries)
- smoke: lsd --tree --max-shown -1 --tree-columns ./src packs horizontally with same cap
- smoke: positive and unset paths unchanged

+++ src/config_file.rs
@@ -44,7 +44,7 @@ pub struct Config {
     pub header: Option<bool>,
     pub literal: Option<bool>,
     pub truncate_owner: Option<TruncateOwner>,
-    pub max_shown: Option<usize>,
+    pub max_shown: Option<i64>,
     pub tree_filter: Option<Vec<String>>,
     pub tree_columns: Option<bool>,
 }
